requirements.txt
moviepy==1.0.3
openai-whisper==20231117
transformers==4.36.2
torch==2.2.1
torchaudio==2.2.1
numpy==1.24.3
opencv-python==4.8.1.78
coqui-tts==0.26.0 